moved Game to pages
